requirements.txt
appdirs==1.4.4
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docutils==0.21.2
fonttools==4.55.3
imageio==2.36.1
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pynrrd==1.1.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.25.0
scikit-learn==1.6.0
scikit-sound==0.3.1
scipy==1.15.0
six==1.17.0
statsmodels==0.14.4
threadpoolctl==3.5.0
tifffile==2024.12.12
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
vtk==9.4.1
